requirements.txt
python-bidi>=0.4.2
Pillow>=9.0.0
numpy>=1.21.0